Feat/token based organization verification

+++ SterCore/Repository/LeaveAllocationRepository.cs
@@ -1,5 +1,6 @@
 ﻿using leave_management.Contracts;
 using leave_management.Data;
+using leave_management.Services.Components.ORI;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
@@ -11,58 +12,93 @@ namespace leave_management.Repository
     public class LeaveAllocationRepository : ILeaveAllocationRepository
     {
         private readonly ApplicationDbContext _db;
+        private readonly IOrganizationResourceManager _organizationManager;
 
-        public LeaveAllocationRepository(ApplicationDbContext db)
+        public LeaveAllocationRepository(ApplicationDbContext db, IOrganizationResourceManager organizationManager)
         {
             _db = db;
+            _organizationManager = organizationManager;
         }
 
         public async Task<bool> CheckAllocation(int leavetypeid, string employeeid)
         {
+            
             var period = DateTime.Now.Year;
+
+            //ORI filtering in find all 
             var leaveAllocations =  await FindAll();
-                return leaveAllocations.Where(q => q.EmployeeId == employeeid && q.LeaveTypeId == leavetypeid && q.Period == period).Any();
+
+            return leaveAllocations
+                .Where(q => q.EmployeeId == employeeid && q.LeaveTypeId == leavetypeid && q.Period == period).Any();
         }
 
         public async Task<bool> Create(LeaveAllocations entity)
         {
+            //ORI separating data beetween organizations
+            entity.OrganizationToken = _organizationManager.GetOrganizationToken();
+
             await _db.LeaveAllocations.AddAsync(entity);
             return await Save();
         }
 
         public async Task<bool> Delete(LeaveAllocations entity)
         {
+            //ORI checking if data is from appropirate organization scope
+            if (entity.OrganizationToken != _organizationManager.GetOrganizationToken())
+            {
+                throw new UnauthorizedAccessException();
+            }
+
             _db.LeaveAllocations.Remove(entity);
             return await Save();
         }
 
         public async Task<ICollection<LeaveAllocations>> FindAll()
         {
+            //ORI getting token to find organization scope
+            var organizationToken = _organizationManager.GetOrganizationToken();
+
             var LeaveAllocations = await _db.LeaveAllocations
                 .Include(q => q.LeaveType)
                 .Include(q => q.Employee)
+
+                //ORI Filtring organizations by their tokens to get scope
+                .Where(q => q.OrganizationToken == organizationToken)
                 .ToListAsync();
             return LeaveAllocations;
         }
 
         public async Task<LeaveAllocations> FindById(int id)
         {
+            //ORI getting token to find organization scope
+            var organizationToken = _organizationManager.GetOrganizationToken();
+
             var LeaveAllocation = await _db.LeaveAllocations
                 .Include(q => q.LeaveType)
                 .Include(q => q.Employee)
+
+                //ORI Filtring organizations by their tokens to get scope
+                .Where(q => q.OrganizationToken == organizationToken)
                 .FirstOrDefaultAsync(q => q.Id == id);
+
             return LeaveAllocation;
         }
 
         public async Task<ICollection<LeaveAllocations>> GetLeaveAllocationsByEmployee(string id)
         {
+            //ORI getting token to find organization scope
+            var organizationToken = _organizationManager.GetOrganizationToken();
+
             var period = DateTime.Now.Year;
             var leaveAllocations = await FindAll();
-            return leaveAllocations.Where(q => q.EmployeeId == id && q.Period == period).ToList();
+            return leaveAllocations
+                .Where(q => q.EmployeeId == id && q.Period == period && q.OrganizationToken == organizationToken)
+                .ToList();
         }
 
         public async Task<LeaveAllocations> GetLeaveAllocationsByEmployeeAndType(string id, int leavetypeid)
         {
+            //ORI implemented in find all
             var period = DateTime.Now.Year;
             var allocations = await FindAll();
             return allocations.FirstOrDefault(q => q.EmployeeId == id && q.Period == period && q.LeaveTypeId == leavetypeid);
@@ -70,7 +106,14 @@ namespace leave_management.Repository
 
         public async Task<bool> Exists(int id)
         {
-            var exists = await _db.LeaveAllocations.AnyAsync(q => q.Id == id);
+            //ORI get token 
+            var organizationToken = _organizationManager.GetOrganizationToken(); 
+
+            var exists = await _db.LeaveAllocations
+
+                //ORI Filtring organizations by their tokens to get scope
+                .Where(q => q.OrganizationToken == organizationToken)
+                .AnyAsync(q => q.Id == id);
             return exists;
         }
 
@@ -82,8 +125,20 @@ namespace leave_management.Repository
 
         public async Task<bool> Update(LeaveAllocations entity)
         {
+            //ORI checking if data is from appropirate organization scope
+            if (entity.OrganizationToken != _organizationManager.GetOrganizationToken())
+            {
+                throw new UnauthorizedAccessException();
+            }
+
             _db.LeaveAllocations.Update(entity);
             return await Save();
+        }
+
+        public void SetToken(LeaveAllocations entity)
+        {
+            var token = _organizationManager.GetOrganizationToken();
+            entity.OrganizationToken = token;
         }
     }
 }
